backend/pyproject.toml
[project]
name = "bonsai-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "apscheduler>=3.11.2",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "fastapi-mail>=1.6.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
